LayoutTests/fast/events/touch/touch-fractional-coordinates.html

<!DOCTYPE html>
<script src='../../../resources/js-test.js'></script>
<script src='../../../resources/ui-helper.js'></script>
<style>
#spacer {
  height: 1000px;
  width: 1000px;
}
iframe {
  width: 100px;
  height: 100px;
  border: 0;
  position:relative;
  top:0.5px;
  left:0.5px;
}
#container {
  position: absolute;
  top: 100px;
  left: 10px;
}
#console {
  margin-top: 200px;
}
</style>
<p id='description'></p>
<div id='container'>
  <iframe id=simpleFrame src='resources/frame-touchevent-forwarder.html'></iframe>
</div>
<div id='console'></div>
<div id='spacer'></div>
<script>
  window.jsTestIsAsync = true;

  eventCount = 0;

  var floatPrecision = 0.00001;
  var frameRect = simpleFrame.getBoundingClientRect();
  let expected = [0.5, frameRect.width - 0.5, 0.25];

  window.addEventListener('message', function (event) {
    if (event.data.type == "touchstart") {
      shouldBeCloseTo('event.data.clientX', expected[eventCount], floatPrecision);
      shouldBeCloseTo('event.data.clientY', expected[eventCount], floatPrecision);
      shouldBeCloseTo('event.data.pageX', expected[eventCount], floatPrecision);
      shouldBeCloseTo('event.data.pageY', expected[eventCount], floatPrecision);
      debug('');

      eventCount++;
    }
  });

  async function runTest() {
    if (!eventSender) {
      debug('This test requires eventSender.');
      return;
    }

    debug('Testing fractional touch inside simple iframe');
    await UIHelper.tapAt(frameRect.left, frameRect.top);

    debug('Testing fractional touch inside rotated iframe');
    simpleFrame.style.transform = "rotate(180deg)";
    await UIHelper.tapAt(frameRect.left, frameRect.top);

    debug('Testing fractional touch inside scaled iframe');
    simpleFrame.style.transform = "scale(2)";
    simpleFrame.style.transformOrigin = "top left";
    await UIHelper.tapAt(frameRect.left, frameRect.top);
    finishJSTest();
  };
  runTest();
</script>
